s3: lift client pool from client header
This commit lifts the client pool to its own header and source file pair
and cleans up the includes in 'client.h'.
