+++ src/v/cloud_storage/remote.h
@@ -18,6 +18,7 @@
 #include "model/metadata.h"
 #include "random/simple_time_jitter.h"
 #include "s3/client.h"
+#include "s3/client_pool.h"
 #include "storage/segment_reader.h"
 #include "utils/retry_chain_node.h"
